requirements.txt
streamlit
numpy
pandas
scikit-learn
opencv-python
pillow
tensorflow
joblib
